backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.10.2
pydantic-settings==2.1.0
stripe==7.6.0
openai==1.55.3
python-dotenv==1.0.1
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1
starlette==0.41.3
